Fix types and more appropriately name data sources
