+++ client/src/app/shared/models/data-source.ts
@@ -1,0 +1,6 @@
+import {Observable} from 'rxjs';
+
+export interface DataSource<T> {
+  connect(): Observable<any>;
+  disconnect(): void;
+}
